+++ Editor/SVGImporter.cs
@@ -23,23 +23,25 @@ namespace Unity.VectorGraphics.Editor
     public enum SVGType
     {
         /// <summary>The SVG file will be imported as a tessellated sprite</summary>
-        VectorSprite,
+        VectorSprite = 0,
 
         /// <summary>The SVG file will be imported as a textured sprite</summary>
-        TexturedSprite,
+        TexturedSprite = 1,
 
         /// <summary>The SVG file will be imported as a Texture2D</summary>
-        Texture2D,
+        Texture2D = 2,
 
 #if UNITY_2019_3_OR_NEWER
         /// <summary>The SVG file will be imported as a UIElement Vector Image asset</summary>
-        UIElement
+        UIElement = 3,
 #endif
+        /// <summary>The SVG file will be imported as a tessellated sprite, compatible with the UI canvas masking system</summary>
+        UISVGImage = 4,
     }
 
     /// <summary>The SVG importer class.</summary>
     [Serializable]
-    [ScriptedImporter(2, "svg")]
+    [ScriptedImporter(4, "svg")]
     public class SVGImporter : ScriptedImporter, ISpriteEditorDataProvider
     {
         internal static readonly string k_PackagePath = "Packages/com.unity.vectorgraphics";
@@ -229,6 +231,14 @@ namespace Unity.VectorGraphics.Editor
         }
         [SerializeField] private int m_SampleCount = 4;
 
+        /// <summary>When importing to an SVGImage, preserves the aspect ratio of the generated sprite.</summary>
+        public bool PreserveSVGImageAspect
+        {
+            get { return m_PreserveSVGImageAspect; }
+            set { m_PreserveSVGImageAspect = value; }
+        }
+        [SerializeField] private bool m_PreserveSVGImageAspect;
+
         [SerializeField]
         private SVGSpriteData m_SpriteData = new SVGSpriteData();
         internal SVGSpriteData GetSVGSpriteData() { return m_SpriteData; }
@@ -298,6 +308,10 @@ namespace Unity.VectorGraphics.Editor
                     sprite = BuildSpriteFromGeometry(geometry, rect);
                     GenerateSpriteAsset(ctx, sprite, name);
                     break;
+                case SVGType.UISVGImage:
+                    sprite = BuildSpriteFromGeometry(geometry, rect);
+                    GenerateUGUISpriteAsset(ctx, sprite, name);
+                    break;
                 case SVGType.TexturedSprite:
                     sprite = BuildSpriteFromGeometry(geometry, rect);
                     GenerateTexturedSpriteAsset(ctx, sprite, name);
@@ -332,7 +346,7 @@ namespace Unity.VectorGraphics.Editor
             EditorUtility.SetDirty(this);
         }
 
-        private void GenerateSpriteAsset(AssetImportContext ctx, Sprite sprite, string name)
+        private void PrepareSpriteAsset(Sprite sprite, string name)
         {
             sprite.name = name + "Sprite";
             if (sprite.texture != null)
@@ -350,6 +364,11 @@ namespace Unity.VectorGraphics.Editor
 #endif
 
             sprite.hideFlags = HideFlags.None;
+        }
+
+        private void GenerateSpriteAsset(AssetImportContext ctx, Sprite sprite, string name)
+        {
+            PrepareSpriteAsset(sprite, name);
 
             ctx.AddObjectToAsset("sprite", sprite);
 
@@ -359,6 +378,32 @@ namespace Unity.VectorGraphics.Editor
             var spriteRenderer = gameObject.AddComponent<SpriteRenderer>();
             spriteRenderer.sprite = sprite;
             spriteRenderer.material = mat;
+
+            SetPhysicsShape(sprite);
+
+            if (sprite.texture != null)
+                ctx.AddObjectToAsset("texAtlas", sprite.texture);
+
+            ctx.AddObjectToAsset("gameObject", gameObject);
+            ctx.SetMainObject(gameObject);
+        }
+
+        private void GenerateUGUISpriteAsset(AssetImportContext ctx, Sprite sprite, string name)
+        {
+            PrepareSpriteAsset(sprite, name);
+
+            ctx.AddObjectToAsset("sprite", sprite);
+
+            Material mat = MaterialForSVGSprite(sprite);
+
+            var gameObject = new GameObject(name);
+            gameObject.AddComponent<CanvasRenderer>();
+            gameObject.AddComponent<RectTransform>();
+
+            var svgImage = gameObject.AddComponent<SVGImage>();
+            svgImage.sprite = sprite;
+            svgImage.material = mat;
+            svgImage.preserveAspect = PreserveSVGImageAspect;
 
             SetPhysicsShape(sprite);
 
@@ -516,9 +561,15 @@ namespace Unity.VectorGraphics.Editor
             string path;
             if (sprite.texture != null)
                 // When texture is present, use the VectorGradient shader
-                path = k_PackagePath + "/Runtime/Materials/Unlit_VectorGradient.mat";
+                path = k_PackagePath + "/Runtime/Materials/Unlit_VectorGradient";
             else
-                path = k_PackagePath + "/Runtime/Materials/Unlit_Vector.mat";
+                path = k_PackagePath + "/Runtime/Materials/Unlit_Vector";
+
+            if (SvgType == SVGType.UISVGImage)
+                path += "UI";
+
+            path += ".mat";
+
             return AssetDatabase.LoadAssetAtPath<Material>(path);
         }
 
@@ -589,13 +640,24 @@ namespace Unity.VectorGraphics.Editor
         {
 
             var sprite = asset as Sprite;
-            if (sprite == null)
-            {
-                var go = asset as GameObject;
-                var sr = go != null ? go.GetComponent<SpriteRenderer>() : null;
-                sprite = sr != null ? sr.sprite : null;
-            }
-            return sprite;
+            if (sprite != null)
+                return sprite;
+
+            var go = asset as GameObject;
+
+            // Try with SpriteRenderer
+            var sr = go != null ? go.GetComponent<SpriteRenderer>() : null;
+            sprite = sr != null ? sr.sprite : null;
+            if (sprite != null)
+                return sprite;
+
+            // Try with VectorImage            
+            var si = go != null ? go.GetComponent<SVGImage>() : null;
+            sprite = si != null ? si.sprite : null;
+            if (sprite != null)
+                return sprite;
+
+            return null;
         }
 
         internal static Texture2D GetImportedTexture2D(string assetPath)